requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pg8000==1.30.5
gunicorn==21.2.0
flask-cors==4.0.0
fpdf2==2.7.8